requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
distlib==0.3.8
distro==1.9.0
Django==5.0.6
djangorestframework==3.15.2
exceptiongroup==1.2.1
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
openai==1.35.13
packaging==24.1
pipenv==2024.0.1
platformdirs==4.2.2
playwright==1.48.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.5.0
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
virtualenv==20.26.3
Werkzeug==3.0.3
